requirements.txt
streamlit
streamlit-pills
llama-index
llama-hub
langchain
pypdf
python-dotenv
openai
nltk
llama-index-vector-stores-milvus
pymilvus
llama-index-llms-anthropic
llama-index-llms-replicate